Developer Build 0.1
Added some basic framework to get started making the ray tracing
element.

It will work by sending rays and checking if they are within triangles.
It will do this by splitting the triangle into 4 parts and finding their
volume, adding these values to eachother. If this sum is the same as the
origonal trianlges volume (or very similar) the line is within the
triangle.

It will send 1 ray per pixel. As of now, the rays will stop where they
land, and because of this there will not be much lighting. This will
change in the future.